requirements.txt
streamlit
openai
python-dotenv
SQLAlchemy
pymysql
pandas
plotly
google-cloud-bigquery